# backend/openapi.yaml
openapi: 3.0.3
info:
  title: BANKERCHANGER API
  version: 0.2.0
  description: REST API for the BANKERCHANGER decentralized boxing prediction market.
servers:
  - url: http://localhost:3001
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    oracleKey:
      type: apiKey
      in: header
      name: X-Oracle-Key
  schemas:
    Error:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
    ErrorResponse400:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, description: "Bad request. Invalid query parameters, request body, or format." }
        code: { type: string, example: "INVALID_REQUEST" }
      required: [success, message]
    ErrorResponse401:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, description: "Missing, invalid, or expired authentication credentials." }
        code: { type: string, example: "UNAUTHORIZED" }
      required: [success, message]
...
    ErrorResponse429:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, description: "Rate limit exceeded. Check Retry-After header for reset time." }
        code: { type: string, example: "RATE_LIMIT_EXCEEDED" }
        retryAfter: { type: integer, description: "Seconds until the rate limit window resets." }
      required: [success, message]
    ErrorResponse500:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, description: "Internal server error. Please try again later." }
        code: { type: string, example: "INTERNAL_SERVER_ERROR" }
        requestId: { type: string, description: "Unique identifier for this error instance for debugging." }
      required: [success, message]
...
paths:
...
  /auth/register:
    post:
      summary: Register a new user
      operationId: register
      tags: [Auth]
      x-rateLimit-windowMs: 60000
      x-rateLimit-max: 10
      x-rateLimit-keyBy: ip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:    { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        '201':
          description: Registration successful; verification email sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:   { type: string }
                  message:  { type: string }
        '400':
          description: Invalid email format or password too short
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse400' }
        '409':
          description: Email already registered
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse429' }
        '500':
          description: Internal server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse500' }
  /auth/login:
    post:
      summary: Login with email and password
      operationId: login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:    { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Tokens issued, or tempToken when 2FA is enabled
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:  { type: string }
                  refreshToken: { type: string }
                  tempToken:    { type: string, description: Present only when 2FA is required }
                  requires2FA:  { type: boolean }
        '400':
          description: Missing or invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse400' }
        '401':
          description: Invalid email or password
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse401' }
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse429' }
        '500':
          description: Internal server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse500' }
  /auth/refresh:
    post:
      summary: Exchange a refresh token for a new access token
      operationId: refresh
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: New access and refresh tokens issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:  { type: string }
                  refreshToken: { type: string }
        '400':
          description: Missing or invalid request body
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse400' }
        '401':
          description: Invalid, expired, or revoked refresh token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse401' }
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse429' }
        '500':
          description: Internal server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse500' }
...